requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cleanup==9.0.0
djangorestframework==3.16.1
Markdown==3.10.2
pillow==12.1.0
Pygments==2.19.2
PyJWT==2.11.0
pymdown-extensions==10.20.1
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.5
tzdata==2025.3
